fix : align waveshare usb-can-a protocol with official spec
- extended frames use 4-byte little-endian IDs on TX and RX
  (ParsedFrameMeta.id widened to u32, DATA_FRAME_MAX_SIZE 13 -> 15)
- CanMode values per spec: Silent=0x01, Loopback=0x02 (were swapped)
- settings frame byte 14: auto-retransmit is now configurable via
  Config::auto_retransmit (default on) instead of hardcoded off
- loopback example switched to LoopbackSilent (hot self-test needs no
  bus partner/ACK) and renamed to tokio_serial_loopback_silent

+++ src/adapters/tokio_serial.rs
@@ -209,9 +209,9 @@ impl<P: Protocol> Transport<P> {
     /// Convert a parsed wire frame into a [`CanMessage`].
     fn parsed_to_message(meta: &ParsedFrameMeta, data: &[u8; 8]) -> Option<CanMessage> {
         let can_id = if meta.is_extended {
-            Id::Extended(ExtendedId::new(meta.id as u32)?)
+            Id::Extended(ExtendedId::new(meta.id)?)
         } else {
-            Id::Standard(StandardId::new(meta.id)?)
+            Id::Standard(StandardId::new(meta.id as u16)?)
         };
         CanMessage::new(can_id, &data[..meta.dlc as usize])
     }
